Create OTH/Viite API modules, add default UTF-8 for sbt opts.
