+++ digiroad2-api-oth/src/main/scala/fi/liikennevirasto/digiroad2/PingApi.scala
