+++ digiroad2-api-oth/src/test/scala/fi/liikennevirasto/digiroad2/PingApiSpec.scala
