+++ project/build.scala
@@ -142,6 +142,74 @@ object Digiroad2Build extends Build {
     )
   ) dependsOn(geoJar, oracleJar, viiteJar)
 
+  val Digiroad2ViiteApiName = "digiroad2-api-viite"
+  lazy val viiteApiJar = Project (
+    Digiroad2ViiteApiName,
+    file(Digiroad2ViiteApiName),
+    settings = Defaults.defaultSettings ++ Seq(
+      organization := Organization,
+      name := Digiroad2ViiteApiName,
+      version := Version,
+      scalaVersion := ScalaVersion,
+      resolvers += Classpaths.typesafeReleases,
+      scalacOptions ++= Seq("-unchecked", "-feature"),
+      //      parallelExecution in Test := false,
+      testOptions in Test ++= (
+        if (System.getProperty("digiroad2.nodatabase", "false") == "true") Seq(Tests.Argument("-l"), Tests.Argument("db")) else Seq()),
+      libraryDependencies ++= Seq(
+        "org.scalatra" %% "scalatra" % ScalatraVersion,
+        "org.scalatra" %% "scalatra-json" % ScalatraVersion,
+        "org.json4s"   %% "json4s-jackson" % "3.2.11",
+        "org.scalatest" % "scalatest_2.11" % "2.2.4" % "test",
+        "org.scalatra" %% "scalatra-scalatest" % ScalatraVersion % "test",
+        "org.scalatra" %% "scalatra-auth" % ScalatraVersion,
+        "org.mockito" % "mockito-core" % "1.9.5" % "test",
+        "com.typesafe.akka" %% "akka-testkit" % "2.3.2" % "test",
+        "ch.qos.logback" % "logback-classic" % "1.0.6" % "runtime",
+        "commons-io" % "commons-io" % "2.4",
+        "com.newrelic.agent.java" % "newrelic-api" % "3.1.1",
+        "org.apache.httpcomponents" % "httpclient" % "4.3.3"
+      ),
+      unmanagedResourceDirectories in Compile += baseDirectory.value / "conf" /  env,
+      unmanagedResourceDirectories in Test += baseDirectory.value / "conf" /  testEnv,
+      unmanagedResourceDirectories in Compile += baseDirectory.value / ".." / "conf" /  env
+    )
+  ) dependsOn(geoJar, oracleJar, viiteJar, commonApiJar)
+
+  val Digiroad2OTHApiName = "digiroad2-api-oth"
+  lazy val othApiJar = Project (
+    Digiroad2OTHApiName,
+    file(Digiroad2OTHApiName),
+    settings = Defaults.defaultSettings ++ Seq(
+      organization := Organization,
+      name := Digiroad2OTHApiName,
+      version := Version,
+      scalaVersion := ScalaVersion,
+      resolvers += Classpaths.typesafeReleases,
+      scalacOptions ++= Seq("-unchecked", "-feature"),
+      //      parallelExecution in Test := false,
+      testOptions in Test ++= (
+        if (System.getProperty("digiroad2.nodatabase", "false") == "true") Seq(Tests.Argument("-l"), Tests.Argument("db")) else Seq()),
+      libraryDependencies ++= Seq(
+        "org.scalatra" %% "scalatra" % ScalatraVersion,
+        "org.scalatra" %% "scalatra-json" % ScalatraVersion,
+        "org.json4s"   %% "json4s-jackson" % "3.2.11",
+        "org.scalatest" % "scalatest_2.11" % "2.2.4" % "test",
+        "org.scalatra" %% "scalatra-scalatest" % ScalatraVersion % "test",
+        "org.scalatra" %% "scalatra-auth" % ScalatraVersion,
+        "org.mockito" % "mockito-core" % "1.9.5" % "test",
+        "com.typesafe.akka" %% "akka-testkit" % "2.3.2" % "test",
+        "ch.qos.logback" % "logback-classic" % "1.0.6" % "runtime",
+        "commons-io" % "commons-io" % "2.4",
+        "com.newrelic.agent.java" % "newrelic-api" % "3.1.1",
+        "org.apache.httpcomponents" % "httpclient" % "4.3.3"
+      ),
+      unmanagedResourceDirectories in Compile += baseDirectory.value / "conf" /  env,
+      unmanagedResourceDirectories in Test += baseDirectory.value / "conf" /  testEnv,
+      unmanagedResourceDirectories in Compile += baseDirectory.value / ".." / "conf" /  env
+    )
+  ) dependsOn(geoJar, oracleJar, commonApiJar)
+
   lazy val warProject = Project (
     Digiroad2Name,
     file("."),
@@ -180,7 +248,8 @@ object Digiroad2Build extends Build {
       unmanagedResourceDirectories in Compile += baseDirectory.value / "conf" /  env,
       unmanagedResourceDirectories in Test += baseDirectory.value / "conf" /  testEnv
     )
-  ) dependsOn(geoJar, oracleJar, viiteJar, commonApiJar) aggregate(geoJar, oracleJar, viiteJar, commonApiJar)
+  ) dependsOn(geoJar, oracleJar, viiteJar, commonApiJar, viiteApiJar, othApiJar) aggregate
+    (geoJar, oracleJar, viiteJar, commonApiJar, viiteApiJar, othApiJar)
 
   lazy val gatling = project.in(file("digiroad2-gatling"))
     .enablePlugins(GatlingPlugin)
